refactor(context): reorganize ProblemsContext import paths and create ProblemsContext file
